aggregate adjacent spans in supervision_staff_attribute_sessions (Recidiviz/recidiviz-data#39569)
## Description of the change

I've noticed recently that I've seen what appears to be an explosion of
rows in `supervision_staff_attribute_sessions`, with a ton of duplciate
adjacent rows. I'm not sure why this wasn't a problem before we added
`is_supervision_officer` and `is_supervision_officer_supervisor`, but
this should fix it.

With this change, we go from 13,627,383 rows in the table in staging to
690,579. Having fewer rows also greatly improves the performance of
`insights_caseload_category_sessions`, which has been alerting recently
because it's taking more than 420 seconds. With this change, it took 62
seconds to load to my sandbox.

Sandbox at `dana_staff`. Let me know if you haveideas for a good way to
verify the change- I did a `compare_views` on
`current_staff_materialized`, which showed no diffs, but I think every
other downstream table is going to have expected diffs because so many
more rows are being merged together now.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 2ba1d09753fb1d459e237d43bf264b52cd9c90e8