[Workflows][US_ID][Bug] Revert idType crcFacilityId for CRC Facilities (Recidiviz/recidiviz-data#39635)
## Description of the change

Reverts 2 commits that add rows with `idType = crcFacilityId` for CRC
facilities in `location_record`. There is an id collision when adding
these rows to Firestore so the FE is missing some CRC Facilities in the
search bar. I have [another
PR](https://github.com/Recidiviz/recidiviz-data/pull/39566) up which is
a path forward to finish the feature but it requires additional testing
before merging so I will go ahead and revert this in the meantime.

This has been run in the jovergaag_us_ix_ sandbox

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 286e8a8ed5f34d95a58bc9607a3cc71b76111be5