[Deploy] Create alpha/* branches with each staging alpha deploy (Recidiviz/recidiviz-data#39448)
## Description of the change

For whatever reason, `git` makes it much easier to rebase on top of a
branch than on top of a tag. It's a common workflow now to rebase on top
of the most recent staging deploy so that we can run
`load_views_to_sandbox` in `auto` mode and pick up fewer changes. I
thought it would be nice to just create an `alpha/<version-tag>` branch
with each alpha deploy to make that easier.

People should now be able to:
* Checkout their feature branch
* Run `ccp origin/alpha/<version> && crb` or `git rebase
origin/alpha/<version>` to rebase on the version branch
* Run `load_views_to_sandbox`
* Run `ccp main && crb` or `git rebase main` to rebase back on the tip
of main

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

N/A

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: cfb8b9c1bdf07c50cb31428b14fcaf598af26b91