[Insights][Zero Grants] Querier mods to prepare for daily zero grant metrics (Recidiviz/recidiviz-data#38903)
## Description of the change

Modify the insights querier to handle end date, include_in_outcomes
value, and `prop_period_critical_caseload` in preparation for the new
calculated-daily ZG metrics.

These changes are closely tied to the `supervision_officer_metrics` view
changes in the next PR so if you'd like more context take a look at the
description/sandbox for Recidiviz/recidiviz-data#39211

This is PR 2/3 to switch zero grant metrics to being calculated daily.
Prior PR is Recidiviz/recidiviz-data#38902. Next PR is Recidiviz/recidiviz-data#39211. Changes in this PR are only in
the insights querier and include:
- Get the relevant ZG end date to use in relevant ZG queries
- Reason: ZG metric rows will now be sourced from a table that updates
daily, so the relevant end date is no longer the first of the month, it
can be any day of the month!
- Pulling the include_in_outcomes value from non-ZG rows
- `include_in_outcomes` will start being `null` for ZG rows (similar to
vitals rows)
- Modifying the prop_period_critical_caseload ZG logic to `>= 1.0`
because of some float value inconsistencies
- Reason: existing bug; this change wasn't specific to daily ZG metrics,
but I noticed it while testing the new view
- Updates the `supervision_officer_metrics` fixture json so that the ZG
metrics have end dates that aren't on the first of the month.

We want these querier changes to work well with the existing set up even
though it's prepping for the incoming change to the
`supervision_officer_metrics` view.

I tested the changes via local backend local frontend and all existing
tests pass so I think this will play nicely with our current view
structure!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#35714

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: cc0e25dd03322b61060e584c154935de19669ee8