No warrant in last 7 years criteria for TX ARS (Recidiviz/recidiviz-data#39605)
## Description of the change

Final criteria for TX ARS: "Have no warrant issued during the previous
seven years of the current parole supervision period. This does not
apply to a warrant issued where subsequently an investigation or
administrative review did not sustain a violation."

I went with using TX raw data because I couldn't figure out how to
interpret the "sustain a violation" piece. I also don't think we have
warrants data in normalized state data? This is the table I'll be
talking to TTs about to try to understand what applies as sustaining a
violation:
https://docs.google.com/spreadsheets/d/1dCcmdw4-LIHz1QdWVOax62YXAaiCeZkowGeVJRtvAp4/edit?gid=0#gid=0

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#38046

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 5cf9e26b6b9c7c63552c620222a20eb9ff1b3c21