[US_TX] Create new edge cases TES for contact standards (Recidiviz/recidiviz-data#39584)
## Description of the change

Defines a criteria view that shows spans of time for which supervision
clients
are compliant with edge case contacts. The following edge case are as
follow:

- "A home contact and residence validation are required within 30
calendar days of the initial
office contact for a newly released or transferred client, unless
otherwise specified in
Section VII. For a client on normal reporting status, home contacts are
required and are
based on their assigned TRAS supervision level and caseload type. Two
separate OIMS
contact entries are not required if the home contact and residence
validation occur on the
same day and the client was present."

- "A home contact is required within 30 calendar days of a client`s
change of address."

- "The PO shall conduct a home contact within five business days after a
client is returned
to supervision from any correctional facility, such as a
TDCJ-Correctional Institutions
Division (CID) unit, TDCJ state jail, Intermediate Sanction Facility
(ISF), Substance
Abuse Felony Punishment Facility (SAFPF), or county jail."

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: cebaffa44c7d7bba8bbf686ff100f10206fd9805