Set new allowed upper bound date for sentence v2 entities (Recidiviz/recidiviz-data#39701)
## Description of the change

This changes the upper bound allowed date for the
`projected_completion_date_max_external` field specifically in the state
and normalized state datasets from 2300-01-01 to 2500-01-01. This is
because this is the latest date we may store on a sentence object, and
we have run into examples where this field has legitimate values later
than 2300-01-01.

One example is that in Utah, charges have statutorily set maximum terms
of imprisonment, and when stored in data these are automatically
calculated as running consecutively. This means that if a person is
convicted on 20 counts of a charge that carries a maximum 20 year
sentence, those 20 year sentences will be summed to determine the
maximum release date for the sentence. This puts
`projected_completion_date_max_external` 400 years from the imposed
date, which is past the current upper bound.

[Sandbox
pipeline](https://console.cloud.google.com/dataflow/jobs/us-east1/2025-03-18_13_06_56-2474491666946846495;graphView=0?project=recidiviz-staging&inv=1&invt=AbsY4A)
showing these changes allow constants to be used in UT without
validations failing

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: af16e6c343b3c33c38a2372b8312d47094065ed2