[US_IA][Workflows] Configure Iowa for workflows  (Recidiviz/recidiviz-data#39708)
## Description of the change

Complete the first two steps
[here](https://www.notion.so/recidiviz/How-to-Configure-a-New-State-1b27889f4d1980bd9fadd1dd32a58bfb)
in order to configure Iowa as a new workflows state. Note that Iowa data
is only in staging and this will have to be updated once the data is
promoted to prod.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

[Recidiviz/recidiviz-dashboardsRecidiviz/recidiviz-data#7700](https://github.com/Recidiviz/recidiviz-dashboards/issues/7700)

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 5bd46a1461e96dbb647c37e4dad95585c5d0dee3