[looker] add model to dashboard generation (Recidiviz/recidiviz-data#39575)
## Description of the change

for state person/normalized state person dashboards we currently create
a
[template](https://github.com/Recidiviz/looker/blob/main/dashboards/state/person_details_template.dashboard.lookml)
dashboard that is then extended to create [prod
](https://github.com/Recidiviz/looker/blob/main/dashboards/state/person_details_prod.dashboard.lookml)and
[staging](https://github.com/Recidiviz/looker/blob/main/dashboards/state/person_details_staging.dashboard.lookml)
dashboards that use the prod and staging models respectively.

justine has been in the process of creating a looker staging and a
looker prod project that can be configured to point to two different
commits of the `looker` repo. the exact deployment process to achieve
this is up in the air but there's been talk of having a step to
`s/model: recidiviz-staging/recidiviz-123` on prod deploy. therefore, we
should write the entity dashboards to use model `recidiviz-staging`.

also fix explore name

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#23292

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1c4d3b08cff9e6e1135fd762c7face8011a26217