[Workflows][US_IX] Prefix crc facility ids with 'CRC ' (Recidiviz/recidiviz-data#39566)
## Description of the change

Adds rows in `location_record` that have `idType = crcFacilityId`, these
rows will be identical to the `facilityId` rows other than the id is
prefixed with 'CRC '. We need to add this prefix this because the import
to Firestore fails on rows with colliding ids. We want to keep both rows
(`crcFacilityId` + `facilityId`) in the system in case we ever want to
search by facility in addition to searching by CRC facility. Some of the
normal facilities have a CRC unit in them, so a facility is not always
one or the other, it is sometimes both.

When we change this id, we must also change the facility id listed in
the `resident.metadata.crcFacilities` so that the facility names match
up, otherwise search will be broken.

In order to make this a non-breaking change, I added/duplicated the
existing facility ids that are in the metadata with the duplicate id now
with 'CRC ' prefix. I will come back around and remove the original ids
once the [FE is swapped
over](https://github.com/Recidiviz/recidiviz-dashboards/pull/7582) to
using the new ids in the production FE.

I'm pretty sure this is non-breaking, but I would appreciate another
opinion on this since I missed the firestore id collision issue with the
original PR which was
[reverted](https://github.com/Recidiviz/recidiviz-data/pull/39635).

This has been run in the jovergaag_us_ix_location sandbox.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of
#[7284](https://github.com/Recidiviz/recidiviz-dashboards/issues/7284)

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 192c06587b0cc5ae9e9fa0e2bae5cfa5e54b3f4c