[Ingest] Add RootEntity properties to Ingest Mappings (Recidiviz/recidiviz-data#39709)
## Description of the change

#### What?
This PR surfaces the `RootEntity` type and external ID types for that
root entity from an ingest mapping. It also:

- Checks that the `RootEntity` is a `StatePerson` or `StateStaff` for
all ingest views (in the new test)
- Checks that some external ID types are defined for the root entity for
all ingest views (in the new test)
- Tests that we raise appropriately if that is not the case (as in one
of our dummy ingest mappings for the compiler test)

#### Why?
For Janet, we will need to subset raw data dependencies to a root
entity's external ID (Recidiviz/recidiviz-data#39679, Recidiviz/recidiviz-data#39680). This change will make it
convenient to surface the necessary ID types to look for in raw data
dependencies (and ensure they are there to begin with)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| **Type: Feature** | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Supports (https://github.com/Recidiviz/recidiviz-data/issues/39679,
https://github.com/Recidiviz/recidiviz-data/issues/39680)

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 6fab2846e52703c33d4a0f260e70c36fb042a92e