[Auth][Admin Panel] Prevent "unknown" role from being assigned permissions (Recidiviz/recidiviz-data#39815)
## Description of the change

Prevents the "unknown" role from being assigned default permissions via
both the API and frontend of the admin panel.

- Adds checks to the API endpoints for adding and updating state roles
- When adding the "unknown" role, if any routes or feature variants are
included in the request, returns an error message.
    - When updating the "unknown" role, returns an error message.
- Disables permissions fields in the admin panel frontend when adding or
updating state role default permissions
- When adding a role, if the role name is "unknown," disables the
permissions fields and displays a message explaining that the unknown
role cannot be assigned permissions. The role is still able to be added
but without any routes/feature variants.
- When updating a role, if any of the selected roles are "unknown",
disables the permissions fields and displays a message explaining that
the unknown role cannot be assigned permissions. Also disables the
"Update" button. The role is still able to be removed.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#39508

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 02f808e1f61ae8b76f30865b7db6eaaba3384d9f